@@ -3,12 +3,12 @@
     flat
   >
     <v-card
-      tile
+      rounded="0"
       flat
     >
       <v-toolbar
         :color="isDark ? '#616161' : '#eeeeee'"
-        dense
+        density="default"
       >
         <v-btn
           icon
@@ -17,209 +17,239 @@
           <v-icon>arrow_back</v-icon>
         </v-btn>
 
-        <v-tooltip bottom>
-          <v-btn
-            slot="activator"
-            :disabled="!isAcked(item.status) && !isClosed(item.status)"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="takeAction(item.id, 'open')"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              :disabled="!isAcked(item.status) && !isClosed(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="takeAction(item.id, 'open')"
             >
-              refresh
-            </v-icon>
-          </v-btn>
+              <v-icon
+                v-bind="props"
+                size="20px"
+              >
+                refresh
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Open') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="!isWatched(item.tags)"
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="watchAlert(item.id)"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="!isWatched(item.tags)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="watchAlert(item.id)"
             >
-              visibility
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                visibility
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Watch') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="isWatched(item.tags)"
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="unwatchAlert(item.id)"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="isWatched(item.tags)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="unwatchAlert(item.id)"
             >
-              visibility_off
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                visibility_off
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Unwatch') }}</span>
+
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="!isAcked(item.status)"
-            slot="activator"
-            :disabled="!isOpen(item.status)"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="ackAlert(item.id)"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="!isAcked(item.status)"
+              :disabled="!isOpen(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="ackAlert(item.id)"
             >
-              check
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                check
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Ack') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="isAcked(item.status)"
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="takeAction(item.id, 'unack')"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="isAcked(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="takeAction(item.id, 'unack')"
             >
-              undo
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                undo
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Unack') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="!isShelved(item.status)"
-            slot="activator"
-            :disabled="!isOpen(item.status) && !isAcked(item.status)"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="shelveAlert(item.id)"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="!isShelved(item.status)"
+              :disabled="!isOpen(item.status) && !isAcked(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="shelveAlert(item.id)"
             >
-              schedule
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                schedule
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Shelve') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            v-show="isShelved(item.status)"
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="takeAction(item.id, 'unshelve')"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              v-show="isShelved(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="takeAction(item.id, 'unshelve')"
             >
-              restore
-            </v-icon>
-          </v-btn>
+              <v-icon
+                v-bind="props"
+                size="20px"
+              >
+                restore
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Unshelve') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            slot="activator"
-            :disabled="isClosed(item.status)"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="takeAction(item.id, 'close')"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              :disabled="isClosed(item.status)"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="takeAction(item.id, 'close')"
             >
-              highlight_off
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                highlight_off
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Close') }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-btn
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="deleteAlert(item.id)"
-          >
-            <v-icon
-              size="20px"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="deleteAlert(item.id)"
             >
-              delete
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                delete
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ $t('Delete') }}</span>
         </v-tooltip>
 
         <v-tooltip
           :key="copyIconText"
-          bottom
+          location="bottom"
         >
-          <v-btn
-            slot="activator"
-            icon
-            class="btn--plain px-1 mx-0"
-            @click="clipboardCopy(item)"
-          >
-            <v-icon
-              size="20px"
+          <template #activator="{props}">
+            <v-btn
+              v-bind="props"
+              icon
+              class="btn--plain px-1 mx-0"
+              @click="clipboardCopy(item)"
             >
-              content_copy
-            </v-icon>
-          </v-btn>
+              <v-icon
+                size="20px"
+                v-bind="props"
+              >
+                content_copy
+              </v-icon>
+            </v-btn>
+          </template>
           <span>{{ copyIconText }}</span>
         </v-tooltip>
 
-        <v-tooltip bottom>
-          <v-menu
-            slot="activator"
-            bottom
-            left
-          >
-            <v-btn
-              slot="activator"
-              icon
-              class="btn--plain px-1 mx-0"
+        <v-tooltip location="bottom">
+          <template #activator="{props}">
+            <v-menu
+              location="bottom"
+              start
             >
-              <v-icon>
-                more_vert
-              </v-icon>
-            </v-btn>
-
-            <v-list
-              subheader
-            >
-              <v-subheader>Actions</v-subheader>
-              <v-divider />
-              <v-list-tile
-                v-for="(action, i) in actions"
-                :key="i"
-                @click="takeAction(item.id, action)"
+              <v-btn
+                v-bind="props"
+                icon
+                class="btn--plain px-1 mx-0"
               >
-                <v-list-tile-title>{{ action | splitCaps }}</v-list-tile-title>
-              </v-list-tile>
-            </v-list>
-          </v-menu>
+                <v-icon v-bind="props">
+                  more_vert
+                </v-icon>
+              </v-btn>
+
+              <v-list>
+                <v-list-subheader>Actions</v-list-subheader>
+                <v-divider />
+                <v-list-item
+                  v-for="(action, i) in actions"
+                  :key="i"
+                  @click="takeAction(item.id, action)"
+                >
+                  <v-list-item-title>{{ this.$filters.splitCaps(action) }}</v-list-item-title>
+                </v-list-item>
+              </v-list>
+            </v-menu>
+          </template>
           <span>{{ $t('More') }}</span>
         </v-tooltip>
       </v-toolbar>
@@ -234,7 +264,15 @@
           <v-tab ripple>
             <v-icon>info</v-icon>&nbsp;{{ $t('Details') }}
           </v-tab>
-          <v-tab-item
+          <v-tab ripple>
+            <v-icon>history</v-icon>&nbsp;{{ $t('History') }}
+          </v-tab>
+          <v-tab ripple>
+            <v-icon>assessment</v-icon>&nbsp;{{ $t('Data') }}
+          </v-tab>
+        </v-tabs>
+        <v-window v-model="active">
+          <v-window-item
             :transition="false"
             :reverse-transition="false"
           >
@@ -245,28 +283,27 @@
                 v-for="note in notes"
                 :key="note.id"
                 :value="true"
-                dismissible
+                closable
                 type="info"
                 class="ma-1"
-                @input="deleteNote(item.id, note.id)"
+                @update:model-value="deleteNote(item.id, note.id)"
               >
                 <b>{{ note.user || 'Anonymous' }}</b> {{ $t('addedNoteOn') }}
                 <span v-if="note.updateTime">
                   <b><date-time
                     :value="note.updateTime"
                     format="longDate"
-                  /></b> ({{ note.updateTime | timeago }})<br>
+                  /></b> ({{ this.$filters.timeago(note.updateTime) }})<br>
                 </span>
                 <span v-else>
                   <b><date-time
                     :value="note.createTime"
                     format="longDate"
-                  /></b> ({{ note.createTime | timeago }})<br>
+                  /></b> ({{ this.$filters.timeago(note.createTime) }})<br>
                 </span>
                 <i>{{ note.text }}</i>
               </v-alert>
 
-              <!-- DEPRECATED -->
               <v-alert
                 v-for="note in historyNotes"
                 :key="note.index"
@@ -279,126 +316,85 @@
                   v-if="note.updateTime"
                   :value="note.updateTime"
                   format="longDate"
-                /></b> ({{ note.updateTime | timeago }})<br>
+                /></b> ({{ this.$filters.timeago(note.updateTime) }})<br>
                 <i>{{ note.text }}</i>
               </v-alert>
               <!-- DEPRECATED -->
 
               <v-card-text>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                <table>
+                  <tbody>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('AlertId') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="console-text">{{ item.id }}</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('LastReceiveAlertId') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="console-text">{{ item.lastReceiveId }}</span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('CreateTime') }}
-                      </div>
-                    </div>
-                    <div class="flex xs9 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <date-time
                           v-if="item.createTime"
                           :value="item.createTime"
                           format="longDate"
                         />
-                        ({{ item.createTime | timeago }})
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                        ({{ this.$filters.timeago(item.createTime) }})
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('ReceiveTime') }}
-                      </div>
-                    </div>
-                    <div class="flex xs9 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <date-time
                           v-if="item.receiveTime"
                           :value="item.receiveTime"
                           format="longDate"
                         />
-                        ({{ item.receiveTime | timeago }})
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                        ({{ this.$filters.timeago(item.receiveTime) }})
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('LastReceiveTime') }}
-                      </div>
-                    </div>
-                    <div class="flex xs9 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <date-time
                           v-if="item.lastReceiveTime"
                           :value="item.lastReceiveTime"
                           format="longDate"
                         />
-                        ({{ item.lastReceiveTime | timeago }})
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div
-                  v-if="$config.customer_views"
-                  class="flex xs12 ma-1"
-                >
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                        ({{ this.$filters.timeago(item.lastReceiveTime) }})
+                      </td>
+                    </tr>
+                    <tr v-if="$config.customer_views">
+                      <td class="text-grey">
                         {{ $t('Customer') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
+                      </td>
+                      <td 
                         class="clickable"
                         @click="queryBy('customer', item.customer)"
                       >
                         {{ item.customer }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Service') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span
                           v-for="service in item.service"
                           :key="service"
@@ -406,70 +402,52 @@
                         >
                           <span class="clickable">{{ service }}</span>&nbsp;
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Environment') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        class="clickable"
-                        @click="queryBy('environment', item.environment)"
-                      >
-                        {{ item.environment }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        <span
+                          class="clickable"
+                          @click="queryBy('environment', item.environment)"
+                        >
+                          {{ item.environment }}
+                        </span>
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Resource') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        class="clickable"
-                        @click="queryBy('resource', item.resource)"
-                      >
-                        {{ item.resource }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        <span
+                          class="clickable"
+                          @click="queryBy('resource', item.resource)"
+                        >
+                          {{ item.resource }}
+                        </span>
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Event') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        class="clickable"
-                        @click="queryBy('event', item.event)"
-                      >
-                        {{ item.event }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        <span
+                          class="clickable"
+                          @click="queryBy('event', item.event)"
+                        >
+                          {{ item.event }}
+                        </span>
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Correlate') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span
                           v-for="event in item.correlate"
                           :key="event"
@@ -477,215 +455,137 @@
                         >
                           <span class="clickable">{{ event }}</span>&nbsp;
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Group') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        class="clickable"
-                        @click="queryBy('group', item.group)"
-                      >
-                        {{ item.group }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        <span
+                          class="clickable"
+                          @click="queryBy('group', item.group)"
+                        >
+                          {{ item.group }}
+                        </span>
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Severity') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span :class="['label', 'label-' + item.previousSeverity]">
-                          {{ item.previousSeverity | capitalize }}
+                          {{ this.$filters.capitalize(item.previousSeverity) }}
                         </span>&nbsp;&rarr;&nbsp;
                         <span :class="['label', 'label-' + item.severity]">
-                          {{ item.severity | capitalize }}
+                          {{ this.$filters.capitalize(item.severity) }}
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Status') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="label">
-                          {{ item.status | capitalize }}
+                          {{ this.$filters.capitalize(item.status) }}
                         </span>
                         <span
                           v-if="statusNote && statusNote.user"
-                        >&nbsp;{{ $t('by') }} <b>{{ statusNote.user }}</b> ({{ statusNote.updateTime | timeago }})
+                        >&nbsp;{{ $t('by') }} <b>{{ statusNote.user }}</b> ({{ this.$filters.timeago(statusNote.updateTime) }})
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div
-                  v-if="statusNote && statusNote.user && statusNote.text"
-                  class="flex xs12 ma-1"
-                >
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text" />
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
-                        <v-icon small>
+                      </td>
+                    </tr>
+                    <tr v-if="statusNote && statusNote.user && statusNote.text">
+                      <td/>
+                      <td>
+                        <v-icon size="small">
                           error_outline
                         </v-icon>
                         <i>&nbsp;{{ statusNote.text }}</i>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Value') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         {{ item.value }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Text') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
-                        <span v-html="item.text" />
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        {{ item.text }}
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('TrendIndication') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="label">
-                          {{ item.trendIndication | splitCaps }}
+                          {{ this.$filters.splitCaps(item.trendIndication) }}
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Timeout') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         {{ item.timeout }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Type') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="label">
-                          {{ item.type | splitCaps }}
+                          {{ this.$filters.splitCaps(item.type) }}
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('DuplicateCount') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         {{ item.duplicateCount }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Repeat') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <span class="label">
-                          {{ item.repeat | capitalize }}
+                          {{ this.$filters.capitalize(item.repeat) }}
                         </span>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Origin') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        class="clickable"
-                        @click="queryBy('origin', item.origin)"
-                      >
-                        {{ item.origin }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div class="flex xs12 ma-1">
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
+                      </td>
+                      <td>
+                        <span
+                          class="clickable"
+                          @click="queryBy('origin', item.origin)"
+                        >
+                          {{ item.origin }}
+                        </span>
+                      </td>
+                    </tr>
+                    <tr>
+                      <td class="text-grey">
                         {{ $t('Tags') }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div>
+                      </td>
+                      <td>
                         <v-chip
                           v-for="tag in item.tags"
                           :key="tag"
@@ -693,60 +593,52 @@
                           small
                           @click="queryBy('tags', tag)"
                         >
-                          <v-icon left>
+                          <v-icon start>
                             label
                           </v-icon>{{ tag }}
                         </v-chip>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <div
-                  v-for="(value, attr) in item.attributes"
-                  :key="attr"
-                  class="flex xs12 ma-1"
-                >
-                  <div class="d-flex align-top">
-                    <div class="flex xs3 text-xs-left">
-                      <div class="grey--text">
-                        {{ attr | splitCaps }}
-                      </div>
-                    </div>
-                    <div class="flex xs6 text-xs-left">
-                      <div
-                        v-if="typeof value === 'object'"
-                      >
-                        <span
-                          v-for="v in value"
-                          :key="v"
-                          @click="queryBy(`_.${attr}`, v)"
+                      </td>
+                    </tr>
+                    <tr
+                      v-for="(value, attr) in item.attributes"
+                      :key="attr"                      
+                    >
+                      <td class="text-grey">
+                        {{ this.$filters.splitCaps(attr) }}
+                      </td>
+                      <td>
+                        <div
+                          v-if="typeof value === 'object'"
                         >
-                          <span class="clickable">{{ v }}</span>&nbsp;
-                        </span>
-                      </div>
-                      <div
-                        v-else-if="typeof value === 'string' && (value.includes('http://') || value.includes('https://'))"
-                        class="link-text"
-                        v-html="value"
-                      />
-                      <div
-                        v-else
-                        class="clickable"
-                        @click="queryBy(`_.${attr}`, value)"
-                      >
-                        {{ value }}
-                      </div>
-                    </div>
-                  </div>
-                </div>
+                          <span
+                            v-for="v in value"
+                            :key="v"
+                            @click="queryBy(`_.${attr}`, v)"
+                          >
+                            <span class="clickable">{{ v }}</span>&nbsp;
+                          </span>
+                        </div>
+                        <div
+                          v-else-if="typeof value === 'string' && (value.includes('http://') || value.includes('https://'))"
+                          class="link-text"
+                        >
+                          {{ value }}
+                        </div>
+                        <div
+                          v-else
+                          class="clickable"
+                          @click="queryBy(`_.${attr}`, value)"
+                        >
+                          {{ value }}
+                        </div>
+                      </td>
+                    </tr>
+                  </tbody>
+                </table>
               </v-card-text>
             </v-card>
-          </v-tab-item>
-
-          <v-tab ripple>
-            <v-icon>history</v-icon>&nbsp;{{ $t('History') }}
-          </v-tab>
-          <v-tab-item
+          </v-window-item>
+          <v-window-item
             :transition="false"
             :reverse-transition="false"
           >
@@ -755,71 +647,68 @@
                 :headers="headersByScreenSize"
                 :items="history"
                 item-key="index"
-                :pagination.sync="pagination"
+                v-model:pagination="pagination"
                 sort-icon="arrow_drop_down"
+                item-props
               >
-                <template
-                  slot="items"
-                  slot-scope="props"
+                <template #item="{item}"
                 >
-                  <td class="hidden-sm-and-down">
-                    <span class="console-text">{{ props.item.id | shortId }}</span>
-                  </td>
-                  <td
-                    class="hidden-sm-and-down text-no-wrap"
-                  >
-                    <date-time
-                      :value="props.item.updateTime"
-                      format="mediumDate"
-                    />
-                  </td>
-                  <td
-                    class="hidden-md-and-up text-no-wrap"
-                  >
-                    <date-time
-                      :value="props.item.updateTime"
-                      format="shortTime"
-                    />
-                  </td>
-                  <td class="hidden-sm-and-down">
-                    <span :class="['label', 'label-' + props.item.severity]">
-                      {{ props.item.severity | capitalize }}
-                    </span>
-                  </td>
-                  <td class="hidden-sm-and-down">
-                    <span class="label">
-                      {{ props.item.status | capitalize }}
-                    </span>
-                  </td>
-                  <td class="hidden-sm-and-down">
-                    {{ props.item.timeout | hhmmss }}
-                  </td>
-                  <td>
-                    <span class="label">
-                      {{ props.item.type || 'unknown' | splitCaps }}
-                    </span>
-                  </td>
-                  <td class="hidden-sm-and-down">
-                    {{ props.item.event }}
-                  </td>
-                  <td class="hidden-sm-and-down">
-                    {{ props.item.value }}
-                  </td>
-                  <td>
-                    {{ props.item.user }}
-                  </td>
-                  <td>
-                    {{ props.item.text }}
-                  </td>
+                  <tr>
+                    <td class="hidden-sm-and-down">
+                      <span class="console-text">{{ this.$filters.shortId(item.props.id) }}</span>
+                    </td>
+                    <td
+                      class="hidden-sm-and-down text-no-wrap"
+                    >
+                      <date-time
+                        :value="item.props.updateTime"
+                        format="mediumDate"
+                      />
+                    </td>
+                    <td
+                      class="hidden-md-and-up text-no-wrap"
+                    >
+                      <date-time
+                        :value="item.props.updateTime"
+                        format="shortTime"
+                      />
+                    </td>
+                    <td class="hidden-sm-and-down">
+                      <span :class="['label', 'label-' + item.props.severity]">
+                        {{ this.$filters.capitalize(item.props.severity) }}
+                      </span>
+                    </td>
+                    <td class="hidden-sm-and-down">
+                      <span class="label">
+                        {{ this.$filters.capitalize(item.props.status) }}
+                      </span>
+                    </td>
+                    <td class="hidden-sm-and-down">
+                      {{ this.$filters.hhmmss(item.props.timeout) }}
+                    </td>
+                    <td>
+                      <span class="label">
+                        {{ this.$filters.splitCaps(item.props.type || 'unknown') }}
+                      </span>
+                    </td>
+                    <td class="hidden-sm-and-down">
+                      {{ item.props.event }}
+                    </td>
+                    <td class="hidden-sm-and-down">
+                      {{ item.props.value }}
+                    </td>
+                    <td>
+                      {{ item.props.user }}
+                    </td>
+                    <td>
+                      {{ item.props.text }}
+                    </td>
+                  </tr>
                 </template>
               </v-data-table>
             </div>
-          </v-tab-item>
-
-          <v-tab ripple>
-            <v-icon>assessment</v-icon>&nbsp;{{ $t('Data') }}
-          </v-tab>
-          <v-tab-item
+          </v-window-item>
+          <v-window-item
             :transition="false"
             :reverse-transition="false"
           >
@@ -833,8 +722,8 @@
                 <span class="console-text">{{ item.rawData || 'no raw data' }}</span>
               </v-card-text>
             </v-card>
-          </v-tab-item>
-        </v-tabs>
+          </v-window-item>
+        </v-window>
       </v-card>
 
       <alert-actions
@@ -882,19 +771,19 @@ export default {
       descending: true
     },
     headers: [
-      { text: i18n.t('AlertOrNoteId'), value: 'id', hide: 'smAndDown' },
-      { text: i18n.t('UpdateTime'), value: 'updateTime', hide: 'smAndDown' },
-      { text: i18n.t('Updated'), value: 'updateTime', hide: 'mdAndUp' },
-      { text: i18n.t('Severity'), value: 'severity', hide: 'smAndDown' },
-      { text: i18n.t('Status'), value: 'status', hide: 'smAndDown' },
-      { text: i18n.t('Timeout'), value: 'timeout', hide: 'smAndDown' },
-      { text: i18n.t('Type'), value: 'type' },
-      { text: i18n.t('Event'), value: 'event', hide: 'smAndDown' },
-      { text: i18n.t('Value'), value: 'value', hide: 'smAndDown' },
-      { text: i18n.t('User'), value: 'user' },
-      { text: i18n.t('Text'), value: 'text' }
+      { title: i18n.global.t('AlertOrNoteId'), key: 'id', hide: 'smAndDown' },
+      { title: i18n.global.t('UpdateTime'), key: 'updateTime', hide: 'smAndDown' },
+      { title: i18n.global.t('Updated'), key: 'updateTime', hide: 'mdAndUp' },
+      { title: i18n.global.t('Severity'), key: 'severity', hide: 'smAndDown' },
+      { title: i18n.global.t('Status'), key: 'status', hide: 'smAndDown' },
+      { title: i18n.global.t('Timeout'), key: 'timeout', hide: 'smAndDown' },
+      { title: i18n.global.t('Type'), key: 'type' },
+      { title: i18n.global.t('Event'), key: 'event', hide: 'smAndDown' },
+      { title: i18n.global.t('Value'), key: 'value', hide: 'smAndDown' },
+      { title: i18n.global.t('User'), key: 'user' },
+      { title: i18n.global.t('Text'), key: 'text' }
     ],
-    copyIconText: i18n.t('Copy')
+    copyIconText: i18n.global.t('Copy')
   }),
   computed: {
     isDark() {
@@ -924,7 +813,7 @@ export default {
     },
     headersByScreenSize() {
       return this.headers.filter(
-        h => !h.hide || !this.$vuetify.breakpoint[h.hide]
+        h => !h.hide || !this.$vuetify.display[h.hide]
       )
     },
     ackTimeout() {
@@ -1012,7 +901,7 @@ export default {
         .then(() => this.getNotes(this.id))
     }, 200, {leading: true, trailing: false}),
     deleteAlert: debounce(function(id) {
-      confirm(i18n.t('ConfirmDelete')) &&
+      confirm(i18n.global.t('ConfirmDelete')) &&
         this.$store.dispatch('alerts/deleteAlert', id)
           .then(() => this.$router.push({ name: 'alerts' }))
     }, 200, {leading: true, trailing: false}),
@@ -1023,7 +912,7 @@ export default {
       this.$emit('close')
     },
     clipboardCopy(item) {
-      this.copyIconText = i18n.t('Copied')
+      this.copyIconText = i18n.global.t('Copied')
 
       let renderedText = this.$config.clipboard_template && nunjucks.renderString(this.$config.clipboard_template, item)
 
@@ -1036,7 +925,7 @@ export default {
       document.execCommand('copy')
       document.body.removeChild(textarea)
       setTimeout(() => {
-        this.copyIconText = i18n.t('Copy')
+        this.copyIconText = i18n.global.t('Copy')
       }, 2000)
     }
   }
