@@ -1,3 +1,4 @@
+<!-- eslint-disable vuetify/no-deprecated-components -->
 <template>
   <div>
     <v-dialog
@@ -7,66 +8,72 @@
       <v-form ref="form">
         <v-card>
           <v-card-title>
-            <span class="headline">
+            <span class="text-h5">
               {{ formTitle }}
             </span>
           </v-card-title>
 
           <v-card-text>
             <v-container grid-list-md>
-              <v-layout wrap>
-                <v-flex
-                  xs12
+              <v-row wrap>
+                <v-col
+                  xs="12"
+                  sm="12"
                 >
                   <v-tooltip
                     :key="copyIconText"
-                    right
+                    end
                   >
-                    <v-text-field
-                      v-if="editedItem.key"
-                      slot="activator"
-                      v-model="editedItem.key"
-                      :label="$t('APIKey')"
-                      readonly
-                      monospace
-                      append-icon="content_copy"
-                      @click:append="clipboardCopy(editedItem.key)"
-                    />
+                    <template #activator>
+                      <v-text-field
+                        v-if="editedItem.key"
+                        v-model="editedItem.key"
+                        :label="$t('APIKey')"
+                        readonly
+                        monospace
+                        append-icon="content_copy"
+                        @click:append="clipboardCopy(editedItem.key)"
+                      />
+                    </template>
                     <span>{{ copyIconText }}</span>
                   </v-tooltip>
-                </v-flex>
-                <v-flex
+                </v-col>
+                <v-col
                   v-if="!isAdmin"
-                  xs12
+                  xs="12"
+                  sm="12"
                 >
                   <v-text-field
                     v-model="editedItem.user"
                     :label="$t('User')"
                     readonly
                   />
-                </v-flex>
-                <v-flex
+                </v-col>
+                <v-col
                   v-if="isAdmin"
-                  xs12
+                  xs="12"
+                  sm="12"
                 >
                   <v-select
                     v-model="editedItem.user"
                     :items="users"
                     :label="$t('User')"
                   />
-                </v-flex>
-                <v-flex
+                </v-col>
+                <v-col
                   v-if="$config.customer_views"
-                  xs12
+                  xs="12"
+                  sm="12"
                 >
                   <v-select
                     v-model="editedItem.customer"
                     :items="allowedCustomers"
                     :label="$t('Customer')"
                   />
-                </v-flex>
-                <v-flex
-                  xs12
+                </v-col>
+                <v-col
+                  xs="12"
+                  sm="12"
                 >
                   <v-autocomplete
                     v-model="editedItem.scopes"
@@ -74,74 +81,80 @@
                     :label="$t('Scopes')"
                     chips
                     clearable
-                    solo
+                    variant="solo"
                     multiple
                   >
                     <template
-                      slot="selection"
-                      slot-scope="data"
+                      #selection="data"
                     >
                       <v-chip
-                        :selected="data.selected"
-                        close
+                        :value="data.selected"
+                        closable
                       >
                         <strong>{{ data.item }}</strong>&nbsp;
                         <span>({{ $t('scope') }})</span>
                       </v-chip>
                     </template>
                   </v-autocomplete>
-                </v-flex>
-                <v-flex
-                  xs12
+                </v-col>
+                <v-col
+                  xs="12"
+                  sm="12"
                 >
                   <v-menu
                     v-model="menu"
                     :close-on-content-click="false"
-                    :nudge-right="40"
+                    :offset="40"
                     lazy
                     transition="scale-transition"
-                    offset-y
                     full-width
                     min-width="290px"
                   >
-                    <v-text-field
-                      slot="activator"
-                      v-model="pickerDate"
-                      :label="$t('Expires')"
-                      prepend-icon="event"
-                      readonly
-                    />
-                    <v-date-picker
+                    <template #activator>
+                      <!--TODO: Native input elements are used while v-date-picker is not yet released for Vuetify 3-->
+                      <label v-html="$t('Expires')" for="date"/>
+                      <br/>
+                      <input type="date" v-model="pickerDate" name="date" class="datetime"/>
+                      <!-- <v-text-field
+                        v-model="pickerDate"
+                        :label="$t('Expires')"
+                        prepend-icon="event"
+                        readonly
+                      /> -->
+                    </template>
+                    <!--TODO: Wait until v-date-picker is readded to Vuetify 3?-->
+                    <!-- <v-date-picker
                       v-model="pickerDate"
                       :min="new Date().toISOString().slice(0, 10)"
-                      @input="menu = false"
-                    />
+                      @update:model-value="menu = false"
+                    /> -->
                   </v-menu>
-                </v-flex>
-                <v-flex
-                  xs12
+                </v-col>
+                <v-col
+                  xs="12"
+                  sm="12"
                 >
                   <v-text-field
                     v-model.trim="editedItem.text"
                     label="Comment"
                   />
-                </v-flex>
-              </v-layout>
+                </v-col>
+              </v-row>
             </v-container>
           </v-card-text>
 
           <v-card-actions>
             <v-spacer />
             <v-btn
-              color="blue darken-1"
-              flat
+              color="blue-darken-1"
+              variant="flat"
               @click="close"
             >
               {{ $t('Cancel') }}
             </v-btn>
             <v-btn
-              color="blue darken-1"
-              flat
+              color="blue-darken-1"
+              variant="flat"
               @click="save"
             >
               {{ $t('Save') }}
@@ -152,199 +165,237 @@
     </v-dialog>
 
     <v-card>
-      <v-card-title class="title">
-        {{ $t('APIKeys') }}
-        <v-spacer />
-        <v-btn-toggle
-          v-model="status"
-          class="transparent"
-          multiple
-        >
-          <v-btn
-            value="active"
-            flat
-          >
-            <v-tooltip bottom>
-              <v-icon slot="activator">
-                check_circle
-              </v-icon>
-              <span>{{ $t('Active') }}</span>
-            </v-tooltip>
-          </v-btn>
-          <v-btn
-            value="expired"
-            flat
-          >
-            <v-tooltip bottom>
-              <v-icon slot="activator">
-                error_outline
-              </v-icon>
-              <span>{{ $t('Expired') }}</span>
-            </v-tooltip>
-          </v-btn>
-        </v-btn-toggle>
-        <v-spacer />
-        <v-text-field
-          v-model="search"
-          append-icon="search"
-          :label="$t('Search')"
-          single-line
-          hide-details
-        />
+      <v-card-title class="text-h6">
+        <v-row>
+          <v-col>
+            {{ $t('APIKeys') }}
+          </v-col>
+          <v-spacer />
+          <v-col>
+            <v-btn-toggle
+              v-model="status"
+              class="bg-transparent"
+              multiple
+            >
+              <v-btn
+                value="active"
+                variant="flat"
+              >
+                <v-tooltip location="bottom">
+                  <template #activator="{props}">
+                    <v-icon v-bind="props">
+                      check_circle
+                    </v-icon>
+                  </template>
+                  <span>{{ $t('Active') }}</span>
+                </v-tooltip>
+              </v-btn>
+              <v-btn
+                value="expired"
+                variant="flat"
+              >
+                <v-tooltip location="bottom">
+                  <template #activator="{props}">
+                    <v-icon v-bind="props">
+                      error_outline
+                    </v-icon>
+                  </template>
+                  <span>{{ $t('Expired') }}</span>
+                </v-tooltip>
+              </v-btn>
+            </v-btn-toggle>
+          </v-col>
+          <v-spacer />
+          <v-col sm="5">
+            <v-text-field
+              v-model="search"
+              append-icon="search"
+              :label="$t('Search')"
+              single-line
+              hide-details
+            />
+          </v-col>
+        </v-row>
       </v-card-title>
 
       <v-data-table
         :headers="computedHeaders"
         :items="keys"
-        :rows-per-page-items="rowsPerPageItems"
-        :pagination.sync="pagination"
+        v-model:pagination="pagination"
         class="px-2"
         :search="search"
         :loading="isLoading"
         must-sort
         sort-icon="arrow_drop_down"
+        item-props
       >
-        <template
-          slot="items"
-          slot-scope="props"
-        >
-          <td
-            class="text-no-wrap"
-            monospace
-          >
-            {{ props.item.key }}
-            <v-tooltip
-              :key="copyIconText"
-              top
+        <template #item="{item}">
+          <tr>
+            <td
+              class="text-no-wrap"
+              monospace
             >
-              <v-icon
-                slot="activator"
-                :value="props.item.key"
-                style="font-size: 16px;"
-                @click="clipboardCopy(props.item.key)"
+              {{ item.props.key }}
+              <v-tooltip
+                :key="copyIconText"
+                location="top"
               >
-                content_copy
-              </v-icon>
-              <span>{{ copyIconText }}</span>
-            </v-tooltip>
-          </td>
-          <td>
-            <v-tooltip
-              v-if="!isExpired(props.item.expireTime)"
-              top
-            >
-              <v-icon
-                slot="activator"
-                color="primary"
+                <template #activator="{props}">
+                  <v-icon
+                    v-bind="props"
+                    :value="item.props.key"
+                    style="font-size: 16px;"
+                    @click="clipboardCopy(item.props.key)"
+                  >
+                    content_copy
+                  </v-icon>
+                </template>
+                <span>{{ copyIconText }}</span>
+              </v-tooltip>
+            </td>
+            <td>
+              <v-tooltip
+                v-if="!isExpired(item.props.expireTime)"
+                location="top"
+              >
+                <template #activator="{props}">
+                  <v-icon
+                    v-bind="props"
+                    color="primary"
+                    size="small"
+                  >
+                    check_circle
+                  </v-icon>
+                </template>
+                <span>{{ $t('Active') }}</span>
+              </v-tooltip>
+              <v-tooltip
+                v-if="isExpired(item.props.expireTime)"
+                location="top"
+              >
+                <template #activator="{props}">
+                  <v-icon
+                    v-bind="props"
+                    color="error"
+                    size="small"
+                  >
+                    error_outline
+                  </v-icon>
+                </template>
+                <span>{{ $t('Expired') }}</span>
+              </v-tooltip>
+            </td>
+            <td>{{ item.props.user }}</td>
+            <td>
+              <v-chip
+                v-for="scope in item.props.scopes"
+                :key="scope"
                 small
               >
-                check_circle
-              </v-icon>
-              <span>{{ $t('Active') }}</span>
-            </v-tooltip>
-            <v-tooltip
-              v-if="isExpired(props.item.expireTime)"
-              top
+                <strong>{{ scope }}</strong>&nbsp;
+                <span>({{ $t('scope') }})</span>
+              </v-chip>
+            </td>
+            <td>{{ item.props.text }}</td>
+            <td>
+              <date-time
+                :value="item.props.expireTime"
+                format="mediumDate"
+              />
+            </td>
+            <td
+              class="text-center"
             >
-              <v-icon
-                slot="activator"
+              {{ item.props.count }}
+            </td>
+            <td>{{ this.$filters.timeago(item.props.lastUsedTime) }}</td> 
+            <td
+              v-if="$config.customer_views"
+            >
+              {{ item.props.customer }}
+            </td>
+            <td class="text-no-wrap">
+              <v-btn
+                v-has-perms.disable="'write:keys'"
+                icon
+                class="btn--plain mr-0"
+                @click="editItem(item.props)"
+              >
+                <v-icon
+                  size="small"
+                  color="grey-darken-3"
+                >
+                  edit
+                </v-icon>
+              </v-btn>
+              <v-btn
+                v-has-perms.disable="'admin:keys'"
+                icon
+                class="btn--plain mx-0"
+                @click="deleteItem(item.props)"
+              >
+                <v-icon
+                  size="small"
+                  color="grey-darken-3"
+                >
+                  delete
+                </v-icon>
+              </v-btn>
+              <v-btn
+                v-has-perms.disable="'admin:keys'"
+                :href="`data:text/plain;base64,${toData(item.props)}`"
+                :download="`key_${item.props.id}.json`"
+                icon
+                class="btn--plain mx-0"
+              >
+                <v-icon
+                  size="small"
+                  color="grey-darken-3"
+                >
+                  get_app
+                </v-icon>
+              </v-btn>
+            </td>
+          </tr>
+        </template>
+        <template #no-data>
+          <tr>
+            <td colspan="7">
+              <v-alert
+                :value="true"
                 color="error"
-                small
+                icon="warning"
               >
-                error_outline
-              </v-icon>
-              <span>{{ $t('Expired') }}</span>
-            </v-tooltip>
-          </td>
-          <td>{{ props.item.user }}</td>
-          <td>
-            <v-chip
-              v-for="scope in props.item.scopes"
-              :key="scope"
-              small
-            >
-              <strong>{{ scope }}</strong>&nbsp;
-              <span>({{ $t('scope') }})</span>
-            </v-chip>
-          </td>
-          <td>{{ props.item.text }}</td>
-          <td>
-            <date-time
-              :value="props.item.expireTime"
-              format="mediumDate"
-            />
-          </td>
-          <td
-            class="text-xs-center"
-          >
-            {{ props.item.count }}
-          </td>
-          <td>{{ props.item.lastUsedTime | timeago }}</td>
-          <td
-            v-if="$config.customer_views"
-          >
-            {{ props.item.customer }}
-          </td>
-          <td class="text-no-wrap">
-            <v-btn
-              v-has-perms.disable="'write:keys'"
-              icon
-              class="btn--plain mr-0"
-              @click="editItem(props.item)"
-            >
-              <v-icon
-                small
-                color="grey darken-3"
-              >
-                edit
-              </v-icon>
-            </v-btn>
-            <v-btn
-              v-has-perms.disable="'admin:keys'"
-              icon
-              class="btn--plain mx-0"
-              @click="deleteItem(props.item)"
-            >
-              <v-icon
-                small
-                color="grey darken-3"
-              >
-                delete
-              </v-icon>
-            </v-btn>
-            <v-btn
-              v-has-perms.disable="'admin:keys'"
-              :href="`data:text/plain;base64,${toData(props.item)}`"
-              :download="`key_${props.item.id}.json`"
-              icon
-              class="btn--plain mx-0"
-            >
-              <v-icon
-                small
-                color="grey darken-3"
-              >
-                get_app
-              </v-icon>
-            </v-btn>
-          </td>
+                {{ $t('NoDisplay') }}
+              </v-alert>
+            </td>
+          </tr>
         </template>
-        <template slot="no-data">
-          <v-alert
-            :value="true"
-            color="error"
-            icon="warning"
-          >
-            {{ $t('NoDisplay') }}
-          </v-alert>
+        <template #no-results>
+          <tr>
+            <td colspan="7">
+              <v-alert
+                :value="true"
+                color="error"
+                icon="warning"
+              >
+                {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
+              </v-alert>
+            </td>
+          </tr>
         </template>
-        <v-alert
-          slot="no-results"
-          :value="true"
-          color="error"
-          icon="warning"
-        >
-          {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
-        </v-alert>
+        <template #bottom>
+          <v-data-table-footer       
+            :items-per-page-options="rowsPerPageItems.map(
+              row => {
+                return {
+                  title: row.toString(),
+                  value: row
+                }
+              }
+            )"
+          />
+        </template>
       </v-data-table>
     </v-card>
 
@@ -379,22 +430,22 @@ export default {
     search: '',
     dialog: false,
     headers: [
-      { text: i18n.t('APIKey'), value: 'key', sortable: false },
-      { text: '', value: 'expireTime' },
-      { text: i18n.t('User'), value: 'user' },
-      { text: i18n.t('Scopes'), value: 'scopes' },
-      { text: i18n.t('Description'), value: 'text' },
-      { text: i18n.t('Expires'), value: 'expireTime' },
-      { text: i18n.t('Count'), value: 'count' },
-      { text: i18n.t('LastUsed'), value: 'lastUsedTime' },
-      { text: i18n.t('Customer'), value: 'customer' },
-      { text: i18n.t('Actions'), value: 'name', sortable: false }
+      { title: i18n.global.t('APIKey'), key: 'key', sortable: false },
+      { title: '', key: 'expireTime' },
+      { title: i18n.global.t('User'), key: 'user' },
+      { title: i18n.global.t('Scopes'), key: 'scopes' },
+      { title: i18n.global.t('Description'), key: 'text' },
+      { title: i18n.global.t('Expires'), key: 'expireTime' },
+      { title: i18n.global.t('Count'), key: 'count' },
+      { title: i18n.global.t('LastUsed'), key: 'lastUsedTime' },
+      { title: i18n.global.t('Customer'), key: 'customer' },
+      { title: i18n.global.t('Actions'), key: 'name', sortable: false }
     ],
     editedId: null,
     editedItem: {
       key: '',
       user: vm.editedId ? null : vm.username(),
-      text: '',
+      title: '',
       customer: null,
       scopes: [],
       expireTime: null
@@ -403,16 +454,16 @@ export default {
     pickerDate: vm.defaultExpireTime(),
     defaultItem: {
       user: vm.editedId ? null : vm.username(),
-      text: '',
+      title: '',
       customer: null,
       scopes: [],
       expireTime: null
     },
-    copyIconText: i18n.t('Copy')
+    copyIconText: i18n.global.t('Copy')
   }),
   computed: {
     computedHeaders() {
-      return this.headers.filter(h => !this.$config.customer_views ? h.value != 'customer' : true)
+      return this.headers.filter(h => !this.$config.customer_views ? h.key != 'customer' : true)
     },
     keys() {
       return this.$store.state.keys.keys.filter(k => !this.status || this.status.includes(this.statusFromExpireTime(k)))
@@ -436,7 +487,7 @@ export default {
       return this.$store.state.keys.isLoading
     },
     formTitle() {
-      return !this.editedId ? i18n.t('NewApiKey') : i18n.t('EditApiKey')
+      return !this.editedId ? i18n.global.t('NewApiKey') : i18n.global.t('EditApiKey')
     },
     refresh() {
       return this.$store.state.refresh
@@ -487,7 +538,7 @@ export default {
       this.dialog = true
     },
     deleteItem(item) {
-      confirm(i18n.t('ConfirmDelete')) &&
+      confirm(i18n.global.t('ConfirmDelete')) &&
         this.$store.dispatch('keys/deleteKey', item.id)
     },
     close() {
@@ -505,7 +556,7 @@ export default {
           {
             user: this.editedItem.user,
             scopes: this.editedItem.scopes,
-            text: this.editedItem.text,
+            title: this.editedItem.text,
             expireTime: this.endOfDay(this.pickerDate),
             customer: this.editedItem.customer
           }
@@ -527,7 +578,7 @@ export default {
       return this.isExpired(key.expireTime) ? 'expired' : 'active'
     },
     clipboardCopy(text) {
-      this.copyIconText = i18n.t('Copied')
+      this.copyIconText = i18n.global.t('Copied')
       let textarea = document.createElement('textarea')
       textarea.textContent = text
       document.body.appendChild(textarea)
@@ -535,7 +586,7 @@ export default {
       document.execCommand('copy')
       document.body.removeChild(textarea)
       setTimeout(() => {
-        this.copyIconText = i18n.t('Copy')
+        this.copyIconText = i18n.global.t('Copy')
       }, 2000)
     },
     toData(item) {
@@ -550,5 +601,12 @@ input[monospace],
 td[monospace] {
   font-family: SFMono-Regular, Consolas, Liberation Mono, Menlo, Courier,
     monospace;
+}
+
+.datetime {
+  border-style: solid;
+  border-radius: 5px;
+  padding: 10px;
+  width: 100%;
 }
 </style>
