@@ -1,3 +1,4 @@
+<!-- eslint-disable vuetify/no-deprecated-components -->
 <template>
   <div>
     <v-dialog
@@ -7,16 +8,16 @@
       <v-form ref="form">
         <v-card>
           <v-card-title>
-            <span class="headline">
+            <span class="text-h5">
               {{ formTitle }}
             </span>
           </v-card-title>
 
           <v-card-text>
             <v-container grid-list-md>
-              <v-layout wrap>
-                <v-flex
-                  xs12
+              <v-row wrap>
+                <v-col
+                  xs="12"
                 >
                   <v-text-field
                     v-model.trim="editedItem.match"
@@ -26,41 +27,41 @@
                     :rules="[rules.required]"
                     required
                   />
-                  <v-flex
-                    xs12
+                  <v-col
+                    xs="12"
                   >
                     <v-chip
                       v-show="editedItem.customer"
-                      close
+                      closable
                       @click="editedItem.customer = null"
                     >
                       <strong>{{ editedItem.customer }}</strong>&nbsp;
                       <span>({{ $t('customer') }})</span>
                     </v-chip>
-                  </v-flex>
+                  </v-col>
                   <v-text-field
                     v-model.trim="editedItem.customer"
                     :label="$t('Customer')"
                     :rules="[rules.required]"
                     required
                   />
-                </v-flex>
-              </v-layout>
+                </v-col>
+              </v-row>
             </v-container>
           </v-card-text>
 
           <v-card-actions>
             <v-spacer />
             <v-btn
-              color="blue darken-1"
-              flat
+              color="blue-darken-1"
+              variant="flat"
               @click="close"
             >
               {{ $t('Cancel') }}
             </v-btn>
             <v-btn
-              color="blue darken-1"
-              flat
+              color="blue-darken-1"
+              variant="flat"
               @click="validate"
             >
               {{ $t('Save') }}
@@ -71,7 +72,7 @@
     </v-dialog>
 
     <v-card>
-      <v-card-title class="title">
+      <v-card-title class="text-h6">
         {{ $t('Customers') }}
         <v-spacer />
         <v-text-field
@@ -86,55 +87,53 @@
       <v-data-table
         :headers="headers"
         :items="customers"
-        :rows-per-page-items="rowsPerPageItems"
-        :pagination.sync="pagination"
+        v-model:pagination="pagination"
         class="px-2"
         :search="search"
         :loading="isLoading"
         must-sort
         sort-icon="arrow_drop_down"
       >
-        <template
-          slot="items"
-          slot-scope="props"
-        >
-          <td>{{ props.item.match }}</td>
-          <td>
-            <v-chip>
-              <strong>{{ props.item.customer }}</strong>&nbsp;
-              <span>({{ $t('customer') }})</span>
-            </v-chip>
-          </td>
-          <td class="text-no-wrap">
-            <v-btn
-              v-has-perms.disable="'admin:customers'"
-              icon
-              class="btn--plain mr-0"
-              @click="editItem(props.item)"
-            >
-              <v-icon
-                small
-                color="grey darken-3"
+        <template #item="{item}">
+          <tr>
+            <td>{{ item.props.title.match }}</td>
+            <td>
+              <v-chip>
+                <strong>{{ item.props.title.customer }}</strong>&nbsp;
+                <span>({{ $t('customer') }})</span>
+              </v-chip>
+            </td>
+            <td class="text-no-wrap">
+              <v-btn
+                v-has-perms.disable="'admin:customers'"
+                icon
+                class="btn--plain mr-0"
+                @click="editItem(item.props.title)"
               >
-                edit
-              </v-icon>
-            </v-btn>
-            <v-btn
-              v-has-perms.disable="'admin:customers'"
-              icon
-              class="btn--plain mx-0"
-              @click="deleteItem(props.item)"
-            >
-              <v-icon
-                small
-                color="grey darken-3"
+                <v-icon
+                  size="small"
+                  color="grey-darken-3"
+                >
+                  edit
+                </v-icon>
+              </v-btn>
+              <v-btn
+                v-has-perms.disable="'admin:customers'"
+                icon
+                class="btn--plain mx-0"
+                @click="deleteItem(item.props.title)"
               >
-                delete
-              </v-icon>
-            </v-btn>
-          </td>
+                <v-icon
+                  size="small"
+                  color="grey-darken-3"
+                >
+                  delete
+                </v-icon>
+              </v-btn>
+            </td>
+          </tr>
         </template>
-        <template slot="no-data">
+        <template #no-data>
           <v-alert
             :value="true"
             color="error"
@@ -143,14 +142,27 @@
             {{ $t('NoDisplay') }}
           </v-alert>
         </template>
-        <v-alert
-          slot="no-results"
-          :value="true"
-          color="error"
-          icon="warning"
-        >
-          {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
-        </v-alert>
+        <template #no-results>
+          <v-alert
+            :value="true"
+            color="error"
+            icon="warning"
+          >
+            {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
+          </v-alert>
+        </template>
+        <template #bottom>
+          <v-data-table-footer       
+            :items-per-page-options="rowsPerPageItems.map(
+              row => {
+                return {
+                  title: row.toString(),
+                  value: row
+                }
+              }
+            )"
+          />
+        </template>
       </v-data-table>
     </v-card>
 
@@ -181,9 +193,9 @@ export default {
     search: '',
     dialog: false,
     headers: [
-      { text: i18n.t('LookUp'), value: 'match' },
-      { text: i18n.t('Customer'), value: 'customer' },
-      { text: i18n.t('Actions'), value: 'name', sortable: false }
+      { title: i18n.global.t('LookUp'), key: 'match' },
+      { title: i18n.global.t('Customer'), key: 'customer' },
+      { title: i18n.global.t('Actions'), key: 'name', sortable: false }
     ],
     editedId: null,
     editedItem: {
@@ -195,7 +207,7 @@ export default {
       customer: null
     },
     rules: {
-      required: v => !!v || i18n.t('Required')
+      required: v => !!v || i18n.global.t('Required')
     }
   }),
   computed: {
@@ -206,7 +218,7 @@ export default {
       return this.$store.state.customers.isLoading
     },
     formTitle() {
-      return !this.editedId ? i18n.t('NewCustomer') : i18n.t('EditCustomer')
+      return !this.editedId ? i18n.global.t('NewCustomer') : i18n.global.t('EditCustomer')
     },
     refresh() {
       return this.$store.state.refresh
@@ -233,7 +245,7 @@ export default {
       this.dialog = true
     },
     deleteItem(item) {
-      confirm(i18n.t('ConfirmDelete')) &&
+      confirm(i18n.global.t('ConfirmDelete')) &&
         this.$store.dispatch('customers/deleteCustomer', item.id)
     },
     close() {
@@ -242,13 +254,15 @@ export default {
         this.$refs.form.reset()
         this.editedItem = Object.assign({}, this.defaultItem)
         this.editedId = null
-      }, 300)
+      }, 100)
     },
     validate() {
-      if (this.$refs.form.validate()) {
-        this.$refs.form.resetValidation()
-        this.save()
-      }
+      this.$refs.form.validate().then((status) => {
+        if(status){
+          this.$refs.form.resetValidation()
+          this.save()
+        }
+      })
     },
     save() {
       if (this.editedId) {
