@@ -1,154 +1,189 @@
+<!-- eslint-disable vuetify/no-deprecated-components -->
 <template>
   <v-card>
-    <v-card-title class="title">
-      {{ $t('Heartbeats') }}
-      <v-spacer />
-      <v-btn-toggle
-        v-model="status"
-        class="transparent"
-        multiple
-      >
-        <v-btn
-          value="ok"
-          flat
-        >
-          <v-tooltip bottom>
-            <v-icon slot="activator">
-              check_circle
-            </v-icon>
-            <span>{{ $t('OK') }}</span>
-          </v-tooltip>
-        </v-btn>
-        <v-btn
-          value="slow"
-          flat
-        >
-          <v-tooltip bottom>
-            <v-icon slot="activator">
-              access_time
-            </v-icon>
-            <span>{{ $t('Slow') }}</span>
-          </v-tooltip>
-        </v-btn>
-        <v-btn
-          value="expired"
-          flat
-        >
-          <v-tooltip bottom>
-            <v-icon slot="activator">
-              timer_off
-            </v-icon>
-            <span>{{ $t('Expired') }}</span>
-          </v-tooltip>
-        </v-btn>
-      </v-btn-toggle>
-      <v-spacer />
-      <v-text-field
-        v-model="search"
-        append-icon="search"
-        :label="$t('Search')"
-        single-line
-        hide-details
-      />
+    <v-card-title class="text-h6">
+      <v-row>
+        <v-col>
+          {{ $t('Heartbeats') }}
+        </v-col>
+        <v-spacer />
+        <v-col>
+          <v-btn-toggle
+            v-model="status"
+            class="bg-transparent"
+            multiple
+          >
+            <v-btn
+              value="ok"
+              variant="flat"
+            >
+              <v-tooltip location="bottom">
+                <template #activator="{props}">
+                  <v-icon v-bind="props">
+                    check_circle
+                  </v-icon>
+                </template>
+                <span>{{ $t('OK') }}</span>
+              </v-tooltip>
+            </v-btn>
+            <v-btn
+              value="slow"
+              variant="flat"
+            >
+              <v-tooltip location="bottom">
+                <template #activator="{props}">
+                  <v-icon v-bind="props">
+                    access_time
+                  </v-icon>
+                </template>
+                <span>{{ $t('Slow') }}</span>
+              </v-tooltip>
+            </v-btn>
+            <v-btn
+              value="expired"
+              variant="flat"
+            >
+              <v-tooltip location="bottom">
+                <template #activator="{props}">
+                  <v-icon v-bind="props">
+                    timer_off
+                  </v-icon>
+                </template>
+                <span>{{ $t('Expired') }}</span>
+              </v-tooltip>
+            </v-btn>
+          </v-btn-toggle>
+        </v-col>
+        <v-spacer />
+        <v-col sm="6">
+          <v-text-field
+            v-model="search"
+            append-icon="search"
+            :label="$t('Search')"
+            single-line
+            hide-details
+          />
+        </v-col>
+      </v-row>
     </v-card-title>
-
     <v-data-table
       :headers="computedHeaders"
       :items="heartbeats"
-      :rows-per-page-items="rowsPerPageItems"
-      :pagination.sync="pagination"
+      v-model:pagination="pagination"
       class="px-2"
       :search="search"
       :loading="isLoading"
       must-sort
       sort-icon="arrow_drop_down"
+      item-props
     >
-      <template
-        slot="items"
-        slot-scope="props"
-      >
-        <td>{{ props.item.origin }}</td>
-        <td
-          v-if="$config.customer_views"
-        >
-          {{ props.item.customer }}
-        </td>
-        <td>
-          <v-chip
-            v-for="tag in props.item.tags"
-            :key="tag"
-            label
-            small
+      <template #item="{item}">
+        <tr>
+          <td>{{ item.props.origin }}</td>
+          <td
+            v-if="this.$config.customer_views"
           >
-            <v-icon left>
+            {{ item.props.customer }}
+          </td>
+          <td>
+            <v-chip
+              v-for="tag in item.props.tags"
+              :key="tag"
               label
-            </v-icon>{{ tag }}
-          </v-chip>
-        </td>
-        <td>
-          {{ props.item.attributes }}
-        </td>
-        <td>
-          <date-time
-            :value="props.item.createTime"
-            format="mediumDate"
-          />
-        </td>
-        <td>
-          <date-time
-            :value="props.item.receiveTime"
-            format="mediumDate"
-          />
-        </td>
-        <td>
-          {{ diffTime(props.item.createTime, props.item.receiveTime) }} ms
-        </td>
-        <td
-          class="text-xs-center text-no-wrap"
-        >
-          {{ timeoutLeft(props.item) | hhmmss }}
-        </td>
-        <td>
-          {{ props.item.receiveTime | timeago }}
-        </td>
-        <td>
-          <span :class="['label', 'label-' + props.item.status.toLowerCase()]">
-            {{ props.item.status | capitalize }}
-          </span>
-        </td>
-        <td class="text-no-wrap">
-          <v-btn
-            v-has-perms.disable="'write:heartbeats'"
-            icon
-            class="btn--plain mr-0"
-            @click="deleteItem(props.item)"
-          >
-            <v-icon
               small
-              color="grey darken-3"
             >
-              delete
-            </v-icon>
-          </v-btn>
-        </td>
+              <v-icon start>
+                label
+              </v-icon>{{ tag }}
+            </v-chip>
+          </td>
+          <td>
+            {{ item.props.attributes }}
+          </td>
+          <td>
+            <date-time
+              :value="item.props.createTime"
+              format="mediumDate"
+            />
+          </td>
+          <td>
+            <date-time
+              :value="item.props.receiveTime"
+              format="mediumDate"
+            />
+          </td>
+          <td>
+            {{ diffTime(item.props.createTime, item.props.receiveTime) }} ms
+          </td>
+          <td
+            class="text-center text-no-wrap"
+          >
+            {{ this.$filters.hhmmss(timeoutLeft(item.props.title)) }}
+          </td>
+          <td>
+            {{ this.$filters.timeago(item.props.receiveTime) }}
+          </td>
+          <td>
+            <span :class="['label', 'label-' + item.props.status.toLowerCase()]">
+              {{ this.$filters.capitalize(item.props.status) }}
+            </span>
+          </td>
+          <td class="text-no-wrap">
+            <v-btn
+              v-has-perms.disable="'write:heartbeats'"
+              icon
+              class="btn--plain mr-0"
+              @click="deleteItem(item.props)"
+            >
+              <v-icon
+                size="small"
+                color="grey-darken-3"
+              >
+                delete
+              </v-icon>
+            </v-btn>
+          </td>
+        </tr>
       </template>
-      <template slot="no-data">
-        <v-alert
-          :value="true"
-          color="error"
-          icon="warning"
-        >
-          {{ $t('NoDisplay') }}
-        </v-alert>
+      <template #no-data>
+        <tr>
+          <td colspan="10">
+            <v-alert
+              :value="true"
+              color="error"
+              icon="warning"
+            >
+              {{ $t('NoDisplay') }}
+            </v-alert>
+          </td>
+        </tr>
       </template>
-      <v-alert
-        slot="no-results"
-        :value="true"
-        color="error"
-        icon="warning"
-      >
-        {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
-      </v-alert>
+      <template #no-results>
+        <tr>
+          <td colspan="10">
+            <v-alert
+              :value="true"
+              color="error"
+              icon="warning"
+            >
+              {{ $t('SearchNoResult1') }} "{{ search }}" {{ $t('SearchNoResult2') }}
+            </v-alert>
+          </td>
+        </tr>
+        
+      </template>
+      <template #bottom>
+        <v-data-table-footer       
+          :items-per-page-options="rowsPerPageItems.map(
+            row => {
+              return {
+                title: row.toString(),
+                value: row
+              }
+            }
+          )"
+        />
+      </template>
     </v-data-table>
   </v-card>
 </template>
@@ -175,17 +210,17 @@ export default {
     status: ['ok', 'slow', 'expired'],
     search: '',
     headers: [
-      { text: i18n.t('Origin'), value: 'origin' },
-      { text: i18n.t('Customer'), value: 'customer' },
-      { text: i18n.t('Tags'), value: 'tags' },
-      { text: i18n.t('Attributes'), value: 'attributes' },
-      { text: i18n.t('CreateTime'), value: 'createTime' },
-      { text: i18n.t('ReceiveTime'), value: 'receiveTime' },
-      { text: i18n.t('Latency'), value: 'latency' },
-      { text: i18n.t('Timeout'), value: 'timeout' },
-      { text: i18n.t('Since'), value: 'since' },
-      { text: i18n.t('Status'), value: 'status' },
-      { text: i18n.t('Actions'), value: 'name', sortable: false }
+      { title: i18n.global.t('Origin'), key: 'origin' },
+      { title: i18n.global.t('Customer'), key: 'customer' },
+      { title: i18n.global.t('Tags'), key: 'tags' },
+      { title: i18n.global.t('Attributes'), key: 'attributes' },
+      { title: i18n.global.t('CreateTime'), key: 'createTime' },
+      { title: i18n.global.t('ReceiveTime'), key: 'receiveTime' },
+      { title: i18n.global.t('Latency'), key: 'latency' },
+      { title: i18n.global.t('Timeout'), key: 'timeout' },
+      { title: i18n.global.t('Since'), key: 'since' },
+      { title: i18n.global.t('Status'), key: 'status' },
+      { title: i18n.global.t('Actions'), key: 'name', sortable: false }
     ]
   }),
   computed: {
@@ -193,7 +228,7 @@ export default {
       return this.$store.state.heartbeats.heartbeats.filter(hb => !this.status || this.status.includes(hb.status))
     },
     computedHeaders() {
-      return this.headers.filter(h => !this.$config.customer_views ? h.value != 'customer' : true)
+      return this.headers.filter(h => !this.$config.customer_views ? h.key != 'customer' : true)
     },
     isLoading() {
       return this.$store.state.heartbeats.isLoading
@@ -219,7 +254,7 @@ export default {
       this.$store.dispatch('heartbeats/getHeartbeats')
     },
     deleteItem(item) {
-      confirm(i18n.t('ConfirmDelete')) &&
+      confirm(i18n.global.t('ConfirmDelete')) &&
         this.$store.dispatch('heartbeats/deleteHeartbeat', item.id)
     },
     diffTime(a, b) {
