@@ -1,20 +1,21 @@
+<!-- eslint-disable vuetify/no-deprecated-components -->
 <template>
   <div>
     <v-data-table
       :headers="headers"
       :items="metrics"
       class="px-2"
-      hide-actions
     >
-      <template
-        slot="items"
-        slot-scope="props"
-      >
-        <td>{{ props.item.title }}</td>
-        <td>{{ props.item.type | capitalize }}</td>
-        <td>{{ props.item.group }}.{{ props.item.name }}</td>
-        <td>{{ props.item.value || props.item.count }}</td>
-        <td>{{ avgTime(props.item) }}</td>
+      <!--TODO: This is to remove the "items per page" selection. Maybe find a more elegant solution?-->
+      <template #bottom />
+      <template #item="{item}">
+        <tr>
+          <td>{{ item.raw.title }}</td>
+          <td>{{ this.$filters.capitalize(item.raw.type) }}</td>
+          <td>{{ item.raw.group }}.{{ item.raw.name }}</td>
+          <td>{{ item.raw.value || item.raw.count }}</td>
+          <td>{{ avgTime(item.raw) }}</td>
+        </tr>
       </template>
     </v-data-table>
 
@@ -22,22 +23,22 @@
       :headers="[{sortable: false},{sortable: false},{sortable: false},{sortable: false}]"
       :items="uptime"
       class="px-2"
-      hide-actions
     >
-      <template
-        slot="items"
-        slot-scope="props"
-      >
-        <td>{{ $t('LastUpdate') }}</td>
-        <td>
-          <date-time
-            v-if="props.item.lastTime"
-            :value="props.item.lastTime"
-            format="longDate"
-          />
-        </td>
-        <td>{{ $t('Uptime') }}</td>
-        <td>{{ props.item.uptime / 1000 | days }}</td>
+      <!--TODO: This is to remove the "items per page" selection. Maybe find a more elegant solution?-->
+      <template #bottom />
+      <template #item="{item}">
+        <tr>
+          <td>{{ $t('LastUpdate') }}</td>
+          <td>
+            <date-time
+              v-if="item.raw.lastTime"
+              :value="item.raw.lastTime"
+              format="longDate"
+            />
+          </td>
+          <td>{{ $t('Uptime') }}</td>
+          <td>{{ this.$filters.days(item.raw.uptime / 1000) }}</td>
+        </tr>
       </template>
     </v-data-table>
   </div>
@@ -54,11 +55,11 @@ export default {
   },
   data: () => ({
     headers: [
-      {text: i18n.t('Metric'), value: 'title', sortable: false},
-      {text: i18n.t('Type'), value: 'type', sortable: false},
-      {text: i18n.t('Name'), value: 'name', sortable: false},
-      {text: i18n.t('Value'), value: 'value', sortable: false},
-      {text: i18n.t('AvgTime'), value: 'time', sortable: false},
+      {title: i18n.global.t('Metric'), key: 'title', sortable: false},
+      {title: i18n.global.t('Type'), key: 'type', sortable: false},
+      {title: i18n.global.t('Name'), key: 'name', sortable: false},
+      {title: i18n.global.t('Value'), key: 'value', sortable: false},
+      {title: i18n.global.t('AvgTime'), key: 'time', sortable: false},
     ]
   }),
   computed: {
