@@ -3,183 +3,177 @@
     max-width="350"
   >
     <v-list>
-      <v-list-tile>
-        <v-list-tile-content>
-          <v-list-tile-title>{{ profile.name }}</v-list-tile-title>
-          <v-list-tile-sub-title>
-            <span>
-              <span
-                v-if="profile.preferred_username && !profile.preferred_username.includes('@')"
-              >@</span>{{ profile.preferred_username }}
-            </span>
-          </v-list-tile-sub-title>
-        </v-list-tile-content>
+      <v-list-item>
+        <v-list-item-title>{{ profile.name }}</v-list-item-title>
+        <v-list-item-subtitle>
+          <span>
+            <span
+              v-if="profile.preferred_username && !profile.preferred_username.includes('@')"
+            >@</span>{{ profile.preferred_username }}
+          </span>
+        </v-list-item-subtitle>
 
-        <v-list-tile-action>
+        <v-list-item-action>
           <v-tooltip
             v-if="profile.provider && profile.provider != 'basic'"
-            top
+            location="top"
           >
-            <v-icon slot="activator">
-              {{
-                provider[profile.provider].icon
-              }}
-            </v-icon>
+            <template #activator="{props}">
+              <v-icon v-bind="props">
+                {{
+                  provider[profile.provider].icon
+                }}
+              </v-icon>
+            </template>
             <span>{{ provider[profile.provider].text }}</span>
           </v-tooltip>
           <v-tooltip
             v-else-if="profile.email_verified"
-            top
+            location="top"
           >
-            <v-icon slot="activator">
-              verified_user
-            </v-icon>
+            <template #activator="{props}">
+              <v-icon v-bind="props">
+                verified_user
+              </v-icon>
+            </template>
             <span>({{ $t('EmailVerified') }})</span>
           </v-tooltip>
           <v-tooltip
             v-else
-            top
+            location="top"
           >
-            <v-icon slot="activator">
-              fas fa-user-times
-            </v-icon>
+            <template #activator="{props}">
+              <v-icon v-bind="props">
+                fas fa-user-times
+              </v-icon>
+            </template>
             <span>{{ $t('EmailNotVerified') }}</span>
           </v-tooltip>
-        </v-list-tile-action>
-      </v-list-tile>
+        </v-list-item-action>
+      </v-list-item>
     </v-list>
 
     <v-divider />
 
     <v-list>
-      <v-list-tile
+      <v-list-item
         v-if="$config.customer_views"
       >
-        <v-list-tile-content>
-          <v-list-tile-title>
-            <span
-              v-for="(customer, index) in customers"
-              :key="index"
+        <v-list-item-title>
+          <span
+            v-for="(customer, index) in customers"
+            :key="index"
+          >
+            <v-chip
+              v-if="index < 3"
+              variant="outlined"
+              small
             >
-              <v-chip
-                v-if="index < 3"
-                outline
-                small
-              >
-                <span>{{ customer }}</span>
-              </v-chip>
-              <span
-                v-if="index === 3"
-                class="grey--text caption"
-              >(+{{ customers.length - 1 }} {{ $t('others') }})</span>
-            </span>
-          </v-list-tile-title>
-          <v-list-tile-sub-title>{{ $t('Customers') }}</v-list-tile-sub-title>
-        </v-list-tile-content>
-      </v-list-tile>
+              <span>{{ customer }}</span>
+            </v-chip>
+            <span
+              v-if="index === 3"
+              class="text-grey text-caption"
+            >(+{{ customers.length - 1 }} {{ $t('others') }})</span>
+          </span>
+        </v-list-item-title>
+        <v-list-item-subtitle>{{ $t('Customers') }}</v-list-item-subtitle>
+      </v-list-item>
 
-      <v-list-tile v-if="profile.orgs">
-        <v-list-tile-content>
-          <v-list-tile-title>
-            <span
-              v-for="(org, index) in profile.orgs"
-              :key="index"
+      <v-list-item v-if="profile.orgs">
+        <v-list-item-title>
+          <span
+            v-for="(org, index) in profile.orgs"
+            :key="index"
+          >
+            <v-chip
+              v-if="index < 3"
+              small
             >
-              <v-chip
-                v-if="index < 3"
-                small
-              >
-                <span>{{ org }}</span>
-              </v-chip>
-              <span
-                v-if="index === 3"
-                class="grey--text caption"
-              >(+{{ profile.orgs.length - 1 }} {{ $t('others') }})</span>
-            </span>
-          </v-list-tile-title>
-          <v-list-tile-sub-title>{{ $t('Organizations') }}</v-list-tile-sub-title>
-        </v-list-tile-content>
-      </v-list-tile>
-      <v-list-tile v-if="profile.groups">
-        <v-list-tile-content>
-          <v-list-tile-title>
+              <span>{{ org }}</span>
+            </v-chip>
             <span
-              v-for="(group, index) in profile.groups"
-              :key="index"
+              v-if="index === 3"
+              class="text-grey text-caption"
+            >(+{{ profile.orgs.length - 1 }} {{ $t('others') }})</span>
+          </span>
+        </v-list-item-title>
+        <v-list-item-subtitle>{{ $t('Organizations') }}</v-list-item-subtitle>
+      </v-list-item>
+      <v-list-item v-if="profile.groups">
+        <v-list-item-title>
+          <span
+            v-for="(group, index) in profile.groups"
+            :key="index"
+          >
+            <v-chip
+              v-if="index < 3"
+              small
             >
-              <v-chip
-                v-if="index < 3"
-                small
-              >
-                <span>{{ group }}</span>
-              </v-chip>
-              <span
-                v-if="index === 3"
-                class="grey--text caption"
-              >(+{{ profile.groups.length - 1 }} {{ $t('others') }})</span>
-            </span>
-          </v-list-tile-title>
-          <v-list-tile-sub-title>{{ $t('Groups') }}</v-list-tile-sub-title>
-        </v-list-tile-content>
-      </v-list-tile>
-      <v-list-tile v-if="profile.roles">
-        <v-list-tile-content>
-          <v-list-tile-title>
+              <span>{{ group }}</span>
+            </v-chip>
             <span
-              v-for="(role, index) in profile.roles"
-              :key="index"
+              v-if="index === 3"
+              class="text-grey text-caption"
+            >(+{{ profile.groups.length - 1 }} {{ $t('others') }})</span>
+          </span>
+        </v-list-item-title>
+        <v-list-item-subtitle>{{ $t('Groups') }}</v-list-item-subtitle>
+      </v-list-item>
+      <v-list-item v-if="profile.roles">
+        <v-list-item-title>
+          <span
+            v-for="(role, index) in profile.roles"
+            :key="index"
+          >
+            <v-chip
+              v-if="index < 3"
+              small
             >
-              <v-chip
-                v-if="index < 3"
-                small
-              >
-                <span>{{ role }}</span>
-              </v-chip>
-              <span
-                v-if="index === 3"
-                class="grey--text caption"
-              >(+{{ profile.roles.length - 1 }} {{ $t('others') }})</span>
-            </span>
-          </v-list-tile-title>
-          <v-list-tile-sub-title>{{ $t('Roles') }}</v-list-tile-sub-title>
-        </v-list-tile-content>
-      </v-list-tile>
-      <v-list-tile>
-        <v-list-tile-content>
-          <v-list-tile-title>
+              <span>{{ role }}</span>
+            </v-chip>
             <span
-              v-for="(scope, index) in scopes"
-              :key="index"
+              v-if="index === 3"
+              class="text-grey text-caption"
+            >(+{{ profile.roles.length - 1 }} {{ $t('others') }})</span>
+          </span>
+        </v-list-item-title>
+        <v-list-item-subtitle>{{ $t('Roles') }}</v-list-item-subtitle>
+      </v-list-item>
+      <v-list-item>
+        <v-list-item-title>
+          <span
+            v-for="(scope, index) in scopes"
+            :key="index"
+          >
+            <v-chip
+              v-if="index < 3"
+              small
             >
-              <v-chip
-                v-if="index < 3"
-                small
-              >
-                <span>{{ scope }}</span>
-              </v-chip>
-              <span
-                v-if="index === 3"
-                class="grey--text caption"
-              >(+{{ scopes.length - 1 }} {{ $t('others') }})</span>
-            </span>
-          </v-list-tile-title>
-          <v-list-tile-sub-title>{{ $t('Scopes') }}</v-list-tile-sub-title>
-        </v-list-tile-content>
-      </v-list-tile>
+              <span>{{ scope }}</span>
+            </v-chip>
+            <span
+              v-if="index === 3"
+              class="text-grey text-caption"
+            >(+{{ scopes.length - 1 }} {{ $t('others') }})</span>
+          </span>
+        </v-list-item-title>
+        <v-list-item-subtitle>{{ $t('Scopes') }}</v-list-item-subtitle>
+      </v-list-item>
     </v-list>
 
     <v-card-actions>
       <v-spacer />
 
       <v-btn
-        flat
+        variant="flat"
         @click="close"
       >
         {{ $t('Cancel') }}
       </v-btn>
       <v-btn
         color="primary"
-        flat
+        variant="flat"
         @click="logout()"
       >
         {{ $t('LogOut') }}
