@@ -3,43 +3,43 @@
     grid-list-sm
     fill-height
   >
-    <v-layout
+    <v-row
       align-center
       row
       wrap
     >
-      <v-flex
-        xs12
-        sm8
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       >
         <div v-show="message">
-          <p class="text-xs-center headline font-weight-medium">
+          <p class="text-center text-h5 font-weight-medium">
             {{ $t('Thanks') }} {{ message }}{{ $t('YouCanNowLogin1') }} <a href="/login">
               {{ $t('YouCanNowLogin2') }}
             </a>
           </p>
         </div>
         <div v-show="error">
-          <p class="text-xs-center headline font-weight-medium">
+          <p class="text-center text-h5 font-weight-medium">
             {{ $t('EmailConfirmFailed') }}
             <a href="/">
               {{ $t('TryAgain') }}
             </a>
           </p>
-          <p class="text-xs-center subheading font-weight-medium">
+          <p class="text-center subheading font-weight-medium">
             {{ $t('Error') }}: {{ error }}
           </p>
         </div>
-      </v-flex>
-      <v-flex
-        xs12
-        sm8
+      </v-col>
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       />
-    </v-layout>
+    </v-row>
   </v-container>
 </template>
 
