@@ -3,19 +3,19 @@
     grid-list-sm
     fill-height
   >
-    <v-layout
+    <v-row
       v-show="!sent"
       align-center
       row
       wrap
     >
-      <v-flex
-        xs12
-        sm8
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       >
-        <p class="text-xs-center headline font-weight-medium">
+        <p class="text-center text-h5 font-weight-medium">
           <span>{{ $t('ResetLink') }}</span>
         </p>
         <v-form @submit.prevent="forgot()">
@@ -25,7 +25,7 @@
             type="text"
             :label="$t('Username')"
             prepend-inner-icon="alternate_email"
-            outline
+            variant="outlined"
           />
           <v-btn
             :loading="isSending"
@@ -37,40 +37,40 @@
             {{ $t('Send') }}
           </v-btn>
         </v-form>
-        <div class="text-xs-center">
-          <span class="body-2">
+        <div class="text-center">
+          <span class="text-body-2">
             {{ $t('AlreadyHaveAccount') }}
           </span>
           <v-btn
-            flat
+            variant="flat"
             color="primary"
             to="/login"
           >
             {{ $t('SignIn') }}
           </v-btn>
         </div>
-      </v-flex>
-      <v-flex
-        xs12
-        sm8
+      </v-col>
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       />
-    </v-layout>
+    </v-row>
 
-    <v-layout
+    <v-row
       v-show="sent"
       align-center
       row
       wrap
     >
-      <v-flex
-        xs12
-        sm8
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       >
-        <p class="text-xs-center headline font-weight-medium">
+        <p class="text-center text-h5 font-weight-medium">
           <span>{{ $t('CheckEmail') }}</span>
         </p>
         <v-form>
@@ -80,7 +80,7 @@
             type="text"
             :label="$t('Username')"
             prepend-inner-icon="alternate_email"
-            outline
+            variant="outlined"
             readonly
           />
           <v-btn
@@ -91,26 +91,26 @@
             {{ $t('ReturnSignIn') }}
           </v-btn>
         </v-form>
-        <div class="text-xs-center">
-          <span class="body-2">
+        <div class="text-center">
+          <span class="text-body-2">
             {{ $t('AlreadyHaveAccount') }}
           </span>
           <v-btn
-            flat
+            variant="flat"
             color="primary"
             to="/login"
           >
             {{ $t('SignIn') }}
           </v-btn>
         </div>
-      </v-flex>
-      <v-flex
-        xs12
-        sm8
+      </v-col>
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       />
-    </v-layout>
+    </v-row>
   </v-container>
 </template>
 
@@ -132,7 +132,7 @@ export default {
       this.$store.dispatch('auth/forgot', this.email).then(() => {
         this.$store.dispatch(
           'notifications/success',
-          i18n.t('ResetEmailSent'),
+          i18n.global.t('ResetEmailSent'),
           { root: true }
         )
         this.sent = true
