@@ -3,18 +3,18 @@
     grid-list-sm
     fill-height
   >
-    <v-layout
+    <v-row
       align-center
       row
       wrap
     >
-      <v-flex
-        xs12
-        sm8
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       >
-        <p class="text-xs-center headline font-weight-medium">
+        <p class="text-center text-h5 font-weight-medium">
           <span>{{ $t('ChooseNewPassword') }}</span>
         </p>
         <v-form @submit.prevent="reset()">
@@ -24,7 +24,7 @@
             :type="showPassword ? 'text' : 'password'"
             :label="$t('Password')"
             :append-icon="showPassword ? 'visibility_off' : 'visibility'"
-            outline
+            variant="outlined"
             @click:append="showPassword = !showPassword"
           />
           <v-text-field
@@ -33,7 +33,7 @@
             :type="showPassword ? 'text' : 'password'"
             :label="$t('ConfirmPassword')"
             :append-icon="showPassword ? 'visibility_off' : 'visibility'"
-            outline
+            variant="outlined"
             @click:append="showPassword = !showPassword"
           />
           <v-btn
@@ -44,26 +44,26 @@
             {{ $t('ResetPassword') }}
           </v-btn>
         </v-form>
-        <div class="text-xs-center">
-          <span class="body-2">
+        <div class="text-center">
+          <span class="text-body-2">
             {{ $t('AlreadyHaveAccount') }}
           </span>
           <v-btn
-            flat
+            variant="flat"
             color="primary"
             to="/login"
           >
             {{ $t('SignIn') }}
           </v-btn>
         </div>
-      </v-flex>
-      <v-flex
-        xs12
-        sm8
+      </v-col>
+      <v-col
+        xs="12"
+        sm="8"
         offset-xs0
         offset-sm2
       />
-    </v-layout>
+    </v-row>
   </v-container>
 </template>
 
