@@ -1,11 +1,9 @@
-import Vue from 'vue'
-
-import {store} from '@/main'
+import {store} from '@/main' 
 
 // v-has-perms.disable="write:keys"
 // v-has-perms="admin:users" (hide is default)
 
-export default Vue.directive('has-perms', function (el, binding) {
+export default function (el, binding) {
   let authRequired = store.getters.getConfig('auth_required')
   let allowReadonly = store.getters.getConfig('allow_readonly')
   let readonlyScopes = store.getters.getConfig('readonly_scopes')
@@ -48,4 +46,4 @@ export default Vue.directive('has-perms', function (el, binding) {
       el.style.display = 'none'
     }
   }
-})
+}
