@@ -1,9 +1,7 @@
-import Vue from 'vue'
+//Vue.filter is no longer used in Vue 3
 
-// See https://vuejs.org/v2/guide/filters.html
-
-export default Vue.filter('capitalize', function (value) {
+export default function (value) {
   if (value == null) return ''
   value = value.toString()
   return value.charAt(0).toUpperCase() + value.slice(1)
-})
+}
