@@ -1,7 +1,6 @@
 import moment from 'moment'
-import Vue from 'vue'
 
-export default Vue.filter('date', function (value, mode = 'local', format = 'll') {
+export default function (value, mode = 'local', format = 'll') {
   if (value) {
     if (mode === 'utc') {
       return moment.utc(String(value)).format(format)
@@ -9,4 +8,4 @@ export default Vue.filter('date', function (value, mode = 'local', format = 'll'
       return moment.utc(String(value)).local().format(format)
     }
   }
-})
+}
