@@ -1,7 +1,6 @@
 import moment from 'moment'
-import Vue from 'vue'
 
-export default Vue.filter('hhmmss', function (value) {
+export default function (value) {
   function pad(s) {
     return ('0' + s).slice(-2)
   }
@@ -12,4 +11,4 @@ export default Vue.filter('hhmmss', function (value) {
     let hours = Math.floor(duration.as('h'))
     return `${hours}:${minutes}:${seconds}`
   }
-})
+}
