@@ -1,10 +1,11 @@
 import bootstrap from './services/config'
 
-import Vue from 'vue'
+import {createApp} from 'vue'
 
 import {createStore} from './store'
 import {createRouter} from './router'
 import {sync} from 'vuex-router-sync'
+import VueAxios from 'vue-axios'
 import axios from 'axios'
 import {makeStore} from '@/store/modules/auth.store'
 import {makeInterceptors} from '@/services/api/interceptors'
@@ -12,30 +13,44 @@ import {vueAuth} from '@/services/auth'
 import GoogleAnalytics from '@/plugins/analytics'
 import i18n from '@/plugins/i18n'
 
-import '@/plugins/vuetify'
-import './stylus/main.styl'
+import vuetify from './plugins/vuetify'
+import './sass/main.sass'
 import App from './App.vue'
 
-import '@/directives/hasPerms'
+import hasPerms from '@/directives/hasPerms'
 
-import '@/filters/capitalize'
-import '@/filters/date'
-import '@/filters/days'
-import '@/filters/hhmmss'
-import '@/filters/shortId'
-import '@/filters/splitCaps'
-import '@/filters/timeago'
-import '@/filters/until'
+import capitalize from '@/filters/capitalize'
+import date from '@/filters/date'
+import days from '@/filters/days'
+import hhmmss from '@/filters/hhmmss'
+import shortId from '@/filters/shortId'
+import splitCaps from '@/filters/splitCaps'
+import timeago from '@/filters/timeago'
+import until from '@/filters/until'
 
 export const store = createStore()
 
+const app = createApp(App)
+app.use(VueAxios, axios)
+app.use(i18n)
+app.use(vuetify)
+app.directive('hasPerms', hasPerms)
+app.config.globalProperties.$filters = {
+  capitalize,
+  date,
+  days,
+  hhmmss,
+  shortId,
+  splitCaps,
+  timeago,
+  until
+}
 bootstrap.getConfig().then(config => {
   const router = createRouter(config.base_path)
-
-  Vue.prototype.$config = config
   store.dispatch('updateConfig', config)
   store.dispatch('alerts/setFilter', config.filter)
-  store.registerModule('auth', makeStore(vueAuth(config)))
+
+  store.registerModule('auth', makeStore(vueAuth(app, config)))
   axios.defaults.baseURL = config.endpoint
 
   const interceptors = makeInterceptors(router)
@@ -43,16 +58,16 @@ bootstrap.getConfig().then(config => {
   axios.interceptors.response.use(undefined, interceptors.interceptErrors)
   axios.interceptors.response.use(undefined, interceptors.redirectToLogin)
 
-  Vue.use(GoogleAnalytics, {
+  app.use(router)
+  app.use(store)
+
+  app.config.globalProperties.$config = config
+
+  app.use(GoogleAnalytics, {
     trackingId: config.tracking_id,
     router
   })
   sync(store, router)
-
-  new Vue({
-    router,
-    store,
-    i18n,
-    render: (h: any) => h(App)
-  }).$mount('#app')
+  app.mount('#app')
 })
+export default app
