@@ -1,6 +1,6 @@
-import Vue from 'vue'
 import axios from 'axios'
 import {AxiosRequestConfig, Method} from 'axios'
+import app from '@/main'
 
 const api = {
   get(url: string, config?: AxiosRequestConfig) {
@@ -31,7 +31,8 @@ const api = {
     let t0 = performance.now()
     return axios.request({...config, url, method, data}).then(response => {
       let t1 = performance.now()
-      Vue.prototype.$track('timing_complete', {
+
+      app.config.globalProperties.$track('timing_complete', {
         name: method,
         event_category: 'API',
         event_label: url,
