@@ -1,19 +1,12 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-import VueAxios from 'vue-axios'
 import {VueAuthenticate} from '@alerta/vue-authenticate'
-import axios from 'axios'
-
-Vue.use(Vuex)
-Vue.use(VueAxios, axios)
 
 function getRedirectUri(path: string) {
   return window.location.origin + (path || '')
 }
 
-export function vueAuth(config) {
+export function vueAuth(app, config) {
   let basePath = config.base_path || process.env.BASE_URL
-  return new VueAuthenticate(Vue.prototype.$http, {
+  return new VueAuthenticate(app.config.globalProperties.$http, {
     tokenPath: 'token',
     tokenName: 'token',
     tokenPrefix: '',
