@@ -5,7 +5,7 @@ declare global {
     // tslint:disable no-empty-interface
     interface Element extends VNode {}
     // tslint:disable no-empty-interface
-    interface ElementClass extends Vue {}
+    //interface ElementClass extends Vue {}
     interface IntrinsicElements {
       [elem: string]: any
     }
