@@ -5,16 +5,15 @@
   </div>
 </template>
 
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator'
+<script>
 import Manifest from '@/components/Manifest.vue'
 import Status from '@/components/Status.vue'
 
-@Component({
+
+export default {
   components: {
     Manifest,
     Status
   }
-})
-export default class About extends Vue {}
+}
 </script>
