@@ -4,10 +4,11 @@ import Vue from 'vue'
 import Vuex, {Store} from 'vuex'
 import Vuetify from 'vuetify'
 import i18n from '@/plugins/i18n'
+import app from '@/main'
 
-Vue.config.silent = true
-Vue.use(Vuetify)
-Vue.use(Vuex)
+//app.config.silent = true
+//app.use(Vuetify)
+//app.use(Vuex)
 
 describe('ApiKeyList', () => {
   let store: Store<any>
